Attempt two at conflict fixing
