+++ examples/pagination_manual.rs
@@ -1,7 +1,7 @@
 //! This example shows how manual pagination works. It's what the raw API
 //! returns, but harder to use than an iterator or stream.
 
-use rspotify::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
+use rspotify_async::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
 
 #[tokio::main]
 async fn main() {
