+++ rspotify-sync/Cargo.toml
@@ -1,0 +1,84 @@
+[package]
+authors = [
+    "Ramsay Leung <ramsayleung@gmail.com>",
+    "Mario Ortiz Manero <marioortizmanero@gmail.com>"
+]
+name = "rspotify-sync"
+version = "0.10.0"
+license = "MIT"
+readme = "../README.md"
+description = "Synchronous Spotify API wrapper"
+homepage = "https://github.com/ramsayleung/rspotify"
+repository = "https://github.com/ramsayleung/rspotify"
+keywords = ["spotify", "api"]
+edition = "2018"
+
+[lib]
+path = "../src/lib.rs"
+
+[dependencies]
+rspotify-macros = { path = "../rspotify-macros", version = "0.10.0" }
+rspotify-model = { path = "../rspotify-model", version = "0.10.0" }
+rspotify-http = { path = "../rspotify-http", version = "0.10.0", default-features = false }
+
+base64 = "0.13.0"
+chrono = { version = "0.4.19", features = ["serde", "rustc-serialize"] }
+dotenv = { version = "0.15.0", optional = true }
+getrandom = "0.2.3"
+log = "0.4.14"
+maybe-async = "0.2.6"
+serde = { version = "1.0.130", default-features = false }
+serde_json = "1.0.67"
+thiserror = "1.0.29"
+url = "2.2.2"
+webbrowser = { version = "0.5.5", optional = true }
+
+[dev-dependencies]
+env_logger = { version = "0.9.0", default-features = false }
+
+[features]
+default = ["client-ureq", "ureq-rustls-tls"]
+
+### Client ###
+cli = ["webbrowser"]
+env-file = ["dotenv"]
+
+### HTTP ###
+# Only synchronous clients are listed here.
+client-ureq = ["rspotify-http/client-ureq", "__sync"]
+
+# Passing the TLS features to ureq.
+ureq-rustls-tls = ["rspotify-http/ureq-rustls-tls"]
+
+# Internal features for checking async or sync compilation
+__async = []
+__sync = ["maybe-async/is_sync"]
+
+[package.metadata.docs.rs]
+# Documenting the CLI methods, and working links for `dotenv`
+features = ["cli", "env-file"]
+
+[[example]]
+name = "pagination_sync"
+required-features = ["env-file", "cli"]
+path = "examples/pagination_sync.rs"
+
+[[example]]
+name = "device"
+required-features = ["env-file", "cli"]
+path = "examples/device.rs"
+
+[[example]]
+name = "me"
+required-features = ["env-file", "cli"]
+path = "examples/me.rs"
+
+[[example]]
+name = "search"
+required-features = ["env-file", "cli"]
+path = "examples/search.rs"
+
+[[example]]
+name = "seek_track"
+required-features = ["env-file", "cli"]
+path = "examples/seek_track.rs"
