+++ rspotify-sync/examples/device.rs
@@ -1,4 +1,4 @@
-use rspotify::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
+use rspotify_sync::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
 
 fn main() {
     // You can use any logger for debugging.
@@ -13,8 +13,7 @@ fn main() {
     let url = spotify.get_authorize_url(false).unwrap();
     spotify.prompt_for_token(&url).unwrap();
 
-    match spotify.seek_track(25000, None) {
-        Ok(_) => println!("change to previous playback successful"),
-        Err(_) => eprintln!("change to previous playback failed"),
-    }
+    let devices = spotify.device();
+
+    println!("Request: {:?}", devices);
 }
