+++ rspotify-sync/examples/me.rs
@@ -1,4 +1,4 @@
-use rspotify::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
+use rspotify_sync::{prelude::*, scopes, AuthCodeSpotify, Credentials, OAuth};
 
 fn main() {
     // You can use any logger for debugging.
